requirements.txt
fastapi==0.114.0
uvicorn==0.23.2
jinja2
python-dotenv
httpx
